:rocket: Renamed language store file and updated imports
