+++ frontend/src/routes/about/+page.svelte
@@ -2,7 +2,7 @@
   import AboutInfo from '$lib/components/about/AboutInfo.svelte';
   import ToggledApplicationInfo from '$lib/components/shared/ToggledApplicationInfo.svelte';
   import { applicationRoutes } from '$lib/config/applications';
-  import { languageStore } from '$lib/util/language';
+  import { languageStore } from '$lib/util/languageStore';
   import { setLocale } from '$lib/util/translations';
   import { onMount } from 'svelte';
